Show playlist length at top (closes #172)
